Big reorg of files & structure
