+++ neurodsp/burst/__init__.py
@@ -1,0 +1,3 @@
+
+from .dualthresh import detect_bursts_dual_threshold
+from .utils import compute_burst_stats
