+++ neurodsp/rhythm/__init__.py
@@ -1,0 +1,3 @@
+
+from .lc import lagged_coherence
+from .swm import sliding_window_matching
