+++ neurodsp/rhythm/lc.py
@@ -1,4 +1,4 @@
-"""Estimate the rhythmicity of a neural oscillation by calculating the lagged coherence metric."""
+"""The lagged coherence algorithm for estimating the rhythmicity of a neural signal."""
 
 import numpy as np
 
@@ -8,7 +8,7 @@ from scipy.signal.windows import hann
 ###################################################################################################
 
 def lagged_coherence(sig, f_range, fs, n_cycles=3, f_step=1, return_spectrum=False):
-    """Quantify the rhythmicity of an oscillation using the lagged coherence measure.
+    """Quantify the rhythmicity of an oscillation using lagged coherence.
 
     Parameters
     ----------
