+++ neurodsp/tests/test_burst_dualthresh.py
@@ -1,0 +1,17 @@
+"""Tests for burst detection functions."""
+
+from neurodsp.burst.dualthresh import *
+from neurodsp.burst.dualthresh import _dual_threshold_split, _rmv_short_periods
+
+###################################################################################################
+###################################################################################################
+
+def test_detect_bursts_dual_threshold():
+    pass
+
+## PRIVATE FUNCTION TESTS
+def test_dual_threshold_split():
+    pass
+
+def tests_rmv_short_periods():
+    pass
