+++ neurodsp/tests/test_burst_utils.py
@@ -1,21 +1,9 @@
 """Tests for burst detection functions."""
 
-from neurodsp.burst import *
+from neurodsp.burst.utils import *
 
 ###################################################################################################
 ###################################################################################################
-
-def test_detect_bursts_dual_threshold():
-    pass
-
 
 def test_compute_burst_stats():
-    pass
-
-## PRIVATE FUNCTION TESTS
-
-def test_dual_threshold_split():
-    pass
-
-def tests_rmv_short_periods():
     pass
