+++ neurodsp/tests/test_rhythm_swm.py
@@ -1,0 +1,17 @@
+"""Test the sliding window matching function."""
+
+from neurodsp.rhythm.swm import *
+
+###################################################################################################
+###################################################################################################
+
+def test_sliding_window_matching():
+    pass
+
+## PRIVATE FUNCTIONS
+
+def test_compute_cost():
+    pass
+
+def test_find_new_windowidx():
+    pass
